color matching and theme improvements
