+++ client/src/theme/variants.tsx
@@ -1,0 +1,7 @@
+import { type MantineThemeOverride } from '@mantine/core';
+
+export const variants: MantineThemeOverride = {
+  components: {
+    // Notification: Notification.extend({ classNames: {} }),
+  },
+};
