+++ client/tailwind.config.ts
@@ -4,16 +4,19 @@ import typography from '@tailwindcss/typography';
 import defaults from 'tailwindcss/defaultTheme';
 import colors from 'tailwindcss/colors';
 
+export const font = ['var(--font-default)', ...defaults.fontFamily.sans];
+
 const config: Config = {
   content: [
     './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/theme/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     extend: {
       fontFamily: {
-        sans: ['var(--font-default)', ...defaults.fontFamily.sans],
+        sans: font,
       },
       colors: {
         dwhite: colors.slate[100],
